Fixed the dreaded infinite loop bug. Implemented a few methods for SoZoomTranslate to ensure that it works correctly in all cases.
git-svn-id: https://free-cad.svn.sourceforge.net/svnroot/free-cad/trunk@5036 e8eeb9e2-ec13-0410-a4a9-efa5cf37419d

+++ src/Mod/Sketcher/Gui/ViewProviderSketch.cpp
@@ -991,7 +991,7 @@ bool ViewProviderSketch::isConstraintAtPosition(const Base::Vector3d &constrPos,
 }
 
 Base::Vector3d ViewProviderSketch::seekConstraintPosition(const Base::Vector3d &suggestedPos,
-                                                          const Base::Vector3d &dir, int step,
+                                                          const Base::Vector3d &dir, float step,
                                                           const SoNode *constraint)
 {
     int multiplier = 0;
